+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Changed
 
-- **(breaking)** [#9](https://github.com/embedded-graphics/tinybmp/pull/9) Added support for BMP images saved in top-down row order. A computed field `data_order` is added to the `Header` struct to capture whether the image data is ordered top-down, or the standard bottom-up. 
+- **(breaking)** [#9](https://github.com/embedded-graphics/tinybmp/pull/9) Added support for BMP images saved in top-down row order. A computed field `row_order` is added to the `Header` struct to capture whether the image data is ordered top-down, or the standard bottom-up. 
 
 ## [0.3.0-beta.1] - 2021-05-24